Rename ItemCollection and Item to Menu and Ticket (#487)
This only renames the Python models. The database tables are yet to be renamed.

I'm uncomfortable with how we're using the term "ticket" here — it refers to the stock-keeping unit, while the actual purchase is recorded as a "line item" in an "order". This has potential for future confusion, so maybe it should be renamed again — perhaps "Sku", creating a noun out of an acronym? Left for a future PR.